+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.1.19] - 2026-03-09
+### Added
+- Function to reset the indexing of a vtkStructuredGrid or similar after extract region. (`resetIndexing`)
+- Function to get the direction vectors of a vtkStructuredGrid or similar. (`getVtsDirectionVectors`)
+
 ## [0.1.18] - 2026-03-03
 ### Added
 - Function to resample a vtkStructuredGrid or similar to an image, aligned to the best axis-aligned orientation of the vtsObj. (`filterResampleToImage_BestAligned`)